Moves classes into common lib directory
