+++ bugs.md
@@ -1,0 +1,6 @@
+* error messages during linking
+
+/usr/bin/ld: _ZN2BS10randomUintE: TLS definition in CMakeFiles/biosim4.dir/random.cpp.o section .tbss mismatches non-TLS reference in common/libbiosim4_common.a(dir.cpp.o)
+
+/usr/bin/ld: common/libbiosim4_common.a: error adding symbols: bad value
+
