+++ src/common/column.cpp
