+++ src/common/common.h
@@ -1,0 +1,19 @@
+//#ifndef COMMON_H_INCLUDED//
+//#define COMMON_H_INCLUDED
+// #pragma once
+
+// #include <cmath>
+// #include <cstdint>
+
+// #include "compass.h"
+
+// struct Polar;
+// struct Dir;
+// struct Coord;
+
+
+// #include "coord.h"
+// #include "dir.h"
+// #include "polar.h"
+
+// #endif
