+++ test/test_polar.cpp
@@ -1,4 +1,7 @@
-#include "../src/basicTypes.h"
+#include "../src/common/polar.h"
+#include "../src/common/dir.h"
+#include "../src/common/compass.h"
+
 #include <gtest/gtest.h>
 
 namespace BS {
